+++ QuiltServer/ChangeLog.md
@@ -34,6 +34,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ### Removed
 
+  - Removed Mods
+    * CraftTweaker
 
 ### Added