+++ QuiltServer/ChangeLog.md
@@ -44,6 +44,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
     * Fabric Seasons
     * Qu Enchantments
     * Spell Engine
+  - Added mods
+    * Quad
 
 ## [0.3.0] - 2023-11-17