+++ QuiltServer/ChangeLog.md
@@ -52,6 +52,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
     * Spell Engine
   - Added mods
     * KubeJS
+    * Quad    
   - Added scripts
     * startup.js
     * server.js